Centralize XPath calls
