+++ src/fore.js
@@ -1,102 +1,5 @@
-import {
-    registerCustomXPathFunction,
-    registerXQueryModule
-} from 'fontoxpath';
 import * as fx from "fontoxpath";
 import {XPathUtil} from "./xpath-util";
-
-const XFORMS_NAMESPACE_URI = 'http://www.w3.org/2002/xforms';
-
-/**
- * Implementation of the functionNameResolver passed to FontoXPath to
- * redirect function resolving for unprefixed functions to either the fn or the xf namespace
- */
-function functionNameResolver ({_prefix, localName}, _arity) {
-	switch (localName) {
-		// TODO: put the full XForms library functions set here
-		case 'instance':
-		case 'depends':
-		case 'boolean-from-string':
-			return {namespaceURI: 'http://www.w3.org/2002/xforms', localName};
-		default:
-			return {namespaceURI: 'http://www.w3.org/2005/xpath-functions', localName};
-	}
-}
-
-registerCustomXPathFunction(
-    { namespaceURI: XFORMS_NAMESPACE_URI, localName: 'instance' },
-    ['xs:string?'],
-    'element()?',
-    (dynamicContext, string) => {
-        // Spec: https://www.w3.org/TR/xforms-xpath/#The_XForms_Function_Library#The_instance.28.29_Function
-        // TODO: handle no string passed (null will be passed instead)
-
-        const {formElement} = dynamicContext.currentContext || this;
-
-        // console.log('fnInstance dynamicContext: ', dynamicContext);
-        // console.log('fnInstance string: ', string);
-
-        const instance = formElement.querySelector(`fx-instance[id=${  string  }]`);
-
-        // const def = instance.getInstanceData();
-        if(instance) {
-            const def = instance.getDefaultContext();
-            // console.log('target instance root node: ', def);
-
-            return def;
-        }
-        return null;
-    }
-);
-
-registerCustomXPathFunction(
-    { namespaceURI: XFORMS_NAMESPACE_URI, localName: 'depends' },
-    ['node()*'],
-    'item()?',
-    (dynamicContext, nodes) => {
-        // Spec: https://www.w3.org/TR/xforms-xpath/#The_XForms_Function_Library#The_instance.28.29_Function
-        // TODO: handle no string passed (null will be passed instead)
-
-        const {formElement} = dynamicContext.currentContext || this;
-
-        // console.log('fnInstance dynamicContext: ', dynamicContext);
-        // console.log('depends on : ', nodes[0]);
-
-        return nodes[0];
-    }
-);
-
-// Implement the XForms standard functions here.
-registerXQueryModule(`
-    module namespace xf="${XFORMS_NAMESPACE_URI}";
-
-    declare %public function xf:boolean-from-string($str as xs:string) as xs:boolean {
-        lower-case($str) = "true" or $str = "1"
-    };
-`);
-
-// How to run XQUF:
-/**
-registerXQueryModule(`
-module namespace my-custom-namespace = "my-custom-uri";
-(:~
-	Insert attribute somewhere
-	~:)
-declare %public %updating function my-custom-namespace:do-something ($ele as element()) as xs:boolean {
-	if ($ele/@done) then false() else
-	(insert node
-	attribute done {"true"}
-	into $ele, true())
-};
-`)
-// At some point:
-const contextNode = null;
-const pendingUpdatesAndXdmValue = evaluateUpdatingExpressionSync('ns:do-something(.)', contextNode, null, null, {moduleImports: {'ns': 'my-custom-uri'}})
-
-console.log(pendingUpdatesAndXdmValue.xdmValue); // this is true or false, see function
-
-executePendingUpdateList(pendingUpdatesAndXdmValue.pendingUpdateList, null, null, null);
-*/
 
 export class Fore{
 
@@ -154,104 +57,6 @@ export class Fore{
         return ns[prefix] || null;
     }
 
-    static getInScopeContext(node,ref){
-        let resultNodeset;
-        let parent;
-        if(node.nodeType === Node.ATTRIBUTE_NODE){
-            parent = node.ownerElement;
-        }else{
-            parent = node.parentNode;
-        }
-
-        const repeatItem = parent.closest('fx-repeatitem');
-        if(repeatItem){
-            return repeatItem.nodeset;
-        }
-        const parentBind = parent.closest('[ref]');
-
-        const ownerForm = parent.closest('fx-form');
-        const model = ownerForm.getModel();
-
-        if(parentBind !== null){
-            resultNodeset = parentBind.nodeset;
-        }else if(XPathUtil.isAbsolutePath(ref)){
-            const instanceId = XPathUtil.getInstanceId(ref);
-            resultNodeset = model.getInstance(instanceId).getDefaultContext();
-        }else if(model.getDefaultInstance() !== null){
-            resultNodeset = model.getDefaultInstance().getDefaultContext();
-        }else{
-            return [];
-        }
-        // todo: no support for xforms 'context' yet - see https://github.com/betterFORM/betterFORM/blob/02fd3ec595fa275589185658f3011a2e2e826f4d/core/src/main/java/de/betterform/xml/xforms/XFormsElement.java#L451
-        return resultNodeset;
-    }
-
-    static evaluateXPath (xpath, contextNode, formElement, namespaceResolver) {
-        return fx.evaluateXPath(
-            xpath,
-            contextNode,
-            null,
-            {},
-            'xs:anyType',
-            {
-                currentContext: {formElement},
-                moduleImports: {
-                    xf: 'http://www.w3.org/2002/xforms'
-                },
-                functionNameResolver,
-                namespaceResolver,
-            });
-    }
-
-    static evaluateToFirst (xpath, contextNode, formElement, namespaceResolver) {
-        return fx.evaluateXPathToFirstNode(
-            xpath,
-            contextNode,
-            null,
-            {},
-            {
-                namespaceResolver,
-                defaultFunctionNamespaceURI: 'http://www.w3.org/2002/xforms',
-                moduleImports: {
-                    xf: 'http://www.w3.org/2002/xforms'
-                },
-                currentContext: {formElement}
-            });
-    }
-
-    static evaluateToNodes (xpath, contextNode, formElement, namespaceResolver) {
-        return fx.evaluateXPathToNodes(
-            xpath,
-            contextNode,
-            null,
-            {},
-            {
-                currentContext: {formElement},
-                functionNameResolver,
-                moduleImports: {
-                    xf: 'http://www.w3.org/2002/xforms'
-                },
-                namespaceResolver,
-            });
-    }
-
-
-    static evaluateToBoolean(xpath, contextNode, formElement, namespaceResolver) {
-        return fx.evaluateXPathToBoolean(
-            xpath,
-            contextNode,
-            null,
-            {},
-            {
-                currentContext: {formElement},
-                functionNameResolver,
-                moduleImports: {
-                    xf: 'http://www.w3.org/2002/xforms'
-                },
-                namespaceResolver,
-            });
-    }
-
 	static get XFORMS_NAMESPACE_URI () {
 		return XFORMS_NAMESPACE_URI;
 	}
@@ -305,18 +110,16 @@ export class Fore{
     static async refreshChildren(startElement){
 
         const refreshed = new Promise(((resolve,reject) => {
-            const children = startElement.children;
+            const {children} = startElement;
             if(children){
                 Array.from(children).forEach(element => {
 
                     // todo: later - check for AVTs
                     if (Fore.isUiElement(element.nodeName) && typeof element.refresh === 'function') {
                         element.refresh();
-                    }else{
-                        if(element.nodeName !== 'fx-MODEL'){
+                    }else if(element.nodeName !== 'fx-MODEL'){
                             Fore.refreshChildren(element);
                         }
-                    }
 
                 });
             }
@@ -339,7 +142,7 @@ export class Fore{
                 attrs[i].value = '';
             }
         }
-        const children = n.children;
+        const {children} = n;
         for (let i = 0 ; i< children.length; i++){
             Fore.clear(children[i]);
         }
